more updates
Former-commit-id: 25b209a289e0ce316562ddd836ca7bc5d4f9505c [formerly 35b69ae811f226d2e2a252d9dafa5cba35b585e0]
Former-commit-id: 42e2eed33848ffaf48a2b4749869b97f5ebe0fae

+++ lagom/transform/smooth_filter.py
@@ -34,7 +34,7 @@ def smooth_filter(x, window_length, polyorder, **kwargs):
         smoothed curve data
     
     """
-    x = np.asarray(x, dtype=np.float32)
+    x = np.asarray(x)
     assert x.ndim == 1, 'only a single vector of scalar values is supported'
     out = savgol_filter(x, window_length, polyorder, **kwargs)
     return out
